Archive/backend/requirements.txt
# SkyVoyage Backend Requirements
# Core FastAPI dependencies
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.0
jinja2==3.1.2
aiofiles==23.2.1
aiohttp==3.9.1
redis==5.0.1
pymongo==4.6.0
motor==3.3.2
# joi removed - JavaScript library mismatch
email-validator==2.1.0
phonenumbers==8.13.25
python-dateutil==2.8.2
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.25.2
celery==5.3.4
flower==2.0.1
prometheus-client==0.19.0
structlog==23.2.0
rich==13.7.0
typer==0.9.0
websockets==12.0
